feat: switch to Connect RPC; introduce Buf CLI
After muhc thought, I decided not to build a HTMX web app, and instead I will be attempting to create a Flutter mobile app that will communicate with this Go backend. Since I had to rewrite my server to no longer serve markup, I decided to do a big refactor to use the Connect RPC protocol. This allows the use of HTTP, gRPC, and/or gRPC-web as the API. At this stage I plan on using gRPC for most, if not all, of the communication from the Flutter frontend.

I also decided to adopt the Buf CLI tool into this project. It was developed by the same team as Connect RPC and plays well with it. This will be used for all Protobuf linting, generating, and breaking-change detection.